requirements.txt
africastalking==1.2.9
asgiref==3.8.1
attrs==24.2.0
black==24.8.0
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.3.0
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.0.2
django-cors-headers==4.4.0
django-filter==24.3
django-multiselectfield==0.1.13
django-rest-authtoken==2.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
littleutils==0.2.4
MarkupSafe==3.0.2
mypy-extensions==1.0.0
node==1.2.2
odict==1.9.0
outdated==0.2.2
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
plumber==1.7
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
schema==0.7.7
sqlparse==0.5.1
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
whitenoise==6.8.2
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==7.0
zope.interface==7.0.3
zope.lifecycleevent==5.0
zope.proxy==6.1
